requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
akshare==1.16.79
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
Cython==3.0.12
decorator==5.2.1
et_xmlfile==2.0.0
fastapi==0.115.12
fonttools==4.57.0
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
html5lib==1.1
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
jsonpath==0.82.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lightgbm==4.6.0
lxml==5.3.2
MarkupSafe==3.0.2
matplotlib==3.10.1
mini-racer==0.12.4
msgpack==1.1.0
multidict==6.4.3
narwhals==1.35.0
nest-asyncio==1.6.0
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.2.1
ply==3.11
propcache==0.3.1
protobuf==5.29.4
pyarrow==19.0.1
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-yfinance==0.2.36
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
starlette==0.46.2
streamlit==1.44.1
ta==0.11.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
thriftpy2==0.5.2
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
tushare==1.4.21
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
webencodings==0.5.1
websocket-client==1.8.0
xlrd==2.0.1
yarl==1.20.0
pytest==8.0.0
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
elasticsearch==8.11.1
